* commit 'ab61fb58':
  msm: camera: Modify interrupt handler for CCI
  msm: camera: Return error in case of secure scm call failure
  msm: camera: Correct ref count for secure fd
  msm: camera: Add extra skip after applying flash off operation
  msm: camera: icp: Use icp base while disabling A5
  msm: camera: mem: Remove lock from get_cpu_buf unless mapping
  msm: camera: fd: Fix the order of clearing and handling IRQs
  msm: camera: isp: add acquire hw handler
  msm: camera: uapi: Add new acquire/release HW IOCTLs
  msm: camera: csid: Disable CSI Rx upon fatal errors
  msm: camera: isp: Delay the change in BW during stop
  msm: camera: crm: Fix the size to copy from user in IOCTL
  msm: camera: flash: Remove regulator power up/down from early PCR
  msm: camera: flash: Remove race condition in subdev close
  msm: camera: Changes to accommodate 32-bit Arch
  msm: camera: isp: start CSID with clock rate per sensor
  msm: camera: Fix Dev contect issue on 32bit Kernel
  msm: camera: isp: Add check for rdi only usecases
  msm: camera: Block release if device is active on link
  msm: camera: core: Change log type in deinit sync object
  msm: camera: crm: Unlink devices before clearing the link
  msm: camera: fd: Add mutex protection to frame done processing

+++ drivers/media/platform/msm/camera/cam_jpeg/jpeg_hw/cam_jpeg_hw_mgr.h
@@ -38,7 +38,7 @@
 struct cam_jpeg_process_frame_work_data_t {
 	uint32_t type;
 	void *data;
-	uint64_t request_id;
+	uintptr_t request_id;
 };
 
 /**
@@ -81,7 +81,7 @@ struct cam_jpeg_hw_cfg_req {
 	struct list_head list;
 	struct cam_hw_config_args hw_cfg_args;
 	uint32_t dev_type;
-	int64_t req_id;
+	uintptr_t req_id;
 };
 
 /**
